+++ src/api/v1/timelines.ts
@@ -212,6 +212,7 @@ app.get(
               ),
           ),
         ),
+        // Hide the posts from the muted accounts:
         notInArray(
           posts.accountId,
           db
@@ -230,6 +231,7 @@ app.get(
               ),
             ),
         ),
+        // Hide the posts from the blocked accounts:
         notInArray(
           posts.accountId,
           db
@@ -237,12 +239,60 @@ app.get(
             .from(blocks)
             .where(eq(blocks.accountId, owner.id)),
         ),
+        // Hide the posts from the accounts who blocked the owner:
         notInArray(
           posts.accountId,
           db
             .select({ accountId: blocks.accountId })
             .from(blocks)
             .where(eq(blocks.blockedAccountId, owner.id)),
+        ),
+        // Hide the shared posts from the muted accounts:
+        or(
+          isNull(posts.sharingId),
+          notInArray(
+            posts.sharingId,
+            db
+              .select({ id: posts.id })
+              .from(posts)
+              .innerJoin(mutes, eq(mutes.mutedAccountId, posts.accountId))
+              .where(
+                and(
+                  eq(mutes.accountId, owner.id),
+                  or(
+                    isNull(mutes.duration),
+                    gt(
+                      sql`${mutes.created} + ${mutes.duration}`,
+                      sql`CURRENT_TIMESTAMP`,
+                    ),
+                  ),
+                ),
+              ),
+          ),
+        ),
+        // Hide the shared posts from the blocked accounts:
+        or(
+          isNull(posts.sharingId),
+          notInArray(
+            posts.sharingId,
+            db
+              .select({ id: posts.id })
+              .from(posts)
+              .innerJoin(blocks, eq(blocks.blockedAccountId, posts.accountId))
+              .where(eq(blocks.accountId, owner.id)),
+          ),
+        ),
+        // Hide the shared posts from the accounts who blocked the owner:
+        or(
+          isNull(posts.sharingId),
+          notInArray(
+            posts.sharingId,
+            db
+              .select({ id: posts.id })
+              .from(posts)
+              .innerJoin(blocks, eq(blocks.accountId, posts.accountId))
+              .where(eq(blocks.blockedAccountId, owner.id)),
+          ),
         ),
         query.max_id == null ? undefined : lt(posts.id, query.max_id),
         query.min_id == null ? undefined : gt(posts.id, query.min_id),